% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{domArr_plots}
\alias{domArr_plots}
\title{convert protein GTF format to rectangle blocks for domains}
\usage{
domArr_plots(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{temp}{The input temperature.}
}
\description{
This function converts input gtf format to domain arrangement plots.
}
\examples{
domArr_plots(file.gtf)
}
